js to ts implementation for context
